// src/api/generated/zod/terminal/terminal.zod.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Core | v1
 * OpenAPI spec version: 1.0.0
 */
import * as zod from "zod";

export const getApiTerminalQueryOffsetRegExpTwo = new RegExp("^-?(?:0|[1-9]\\d*)$");
export const getApiTerminalQueryLimitRegExpTwo = new RegExp("^-?(?:0|[1-9]\\d*)$");

export const GetApiTerminalQueryParams = zod.object({
  Search: zod.string().optional(),
  HarborId: zod.uuid().optional(),
  Offset: zod
    .union([zod.int(), zod.stringFormat("int32", getApiTerminalQueryOffsetRegExpTwo)])
    .optional(),
  Limit: zod
    .union([zod.int(), zod.stringFormat("int32", getApiTerminalQueryLimitRegExpTwo)])
    .optional(),
  Sort: zod.string().optional(),
});

export const getApiTerminalResponseTotalRegExpTwo = new RegExp("^-?(?:0|[1-9]\\d*)$");
export const getApiTerminalResponseOffsetRegExpTwo = new RegExp("^-?(?:0|[1-9]\\d*)$");
export const getApiTerminalResponseLimitRegExpTwo = new RegExp("^-?(?:0|[1-9]\\d*)$");

export const GetApiTerminalResponse = zod.object({
  items: zod.array(
    zod.object({
      name: zod.string(),
      harborId: zod.uuid(),
      id: zod.uuid(),
      createdAt: zod.iso.datetime({ offset: true }),
      updatedAt: zod.iso.datetime({ offset: true }),
    }),
  ),
  total: zod.union([zod.int(), zod.stringFormat("int32", getApiTerminalResponseTotalRegExpTwo)]),
  offset: zod.union([zod.int(), zod.stringFormat("int32", getApiTerminalResponseOffsetRegExpTwo)]),
  limit: zod.union([zod.int(), zod.stringFormat("int32", getApiTerminalResponseLimitRegExpTwo)]),
  hasNext: zod.boolean().optional(),
  hasPrevious: zod.boolean().optional(),
});

export const postApiTerminalBodyNameMin = 3;
export const postApiTerminalBodyNameMax = 100;

export const PostApiTerminalBody = zod.object({
  name: zod.string().min(postApiTerminalBodyNameMin).max(postApiTerminalBodyNameMax),
  harborId: zod.uuid(),
});

export const PostApiTerminalResponse = zod.object({
  name: zod.string(),
  harborId: zod.uuid(),
  id: zod.uuid(),
  createdAt: zod.iso.datetime({ offset: true }),
  updatedAt: zod.iso.datetime({ offset: true }),
});

export const GetApiTerminalIdParams = zod.object({
  id: zod.uuid(),
});

export const GetApiTerminalIdResponse = zod.object({
  name: zod.string(),
  harborId: zod.uuid(),
  id: zod.uuid(),
  createdAt: zod.iso.datetime({ offset: true }),
  updatedAt: zod.iso.datetime({ offset: true }),
});

export const PutApiTerminalIdParams = zod.object({
  id: zod.uuid(),
});

export const putApiTerminalIdBodyNameMin = 3;
export const putApiTerminalIdBodyNameMax = 100;

export const PutApiTerminalIdBody = zod.object({
  name: zod.string().min(putApiTerminalIdBodyNameMin).max(putApiTerminalIdBodyNameMax),
  harborId: zod.uuid(),
});

export const PutApiTerminalIdResponse = zod.object({
  name: zod.string(),
  harborId: zod.uuid(),
  id: zod.uuid(),
  createdAt: zod.iso.datetime({ offset: true }),
  updatedAt: zod.iso.datetime({ offset: true }),
});

export const DeleteApiTerminalIdParams = zod.object({
  id: zod.uuid(),
});

export const DeleteApiTerminalIdResponse = zod.unknown();